+++ core/net/rime/chameleon-bitopt.c
@@ -93,7 +93,7 @@ le16_read(const void *ptr)
 }
 /*---------------------------------------------------------------------------*/
 static uint8_t CC_INLINE
-get_bits_in_byte(uint8_t *from, int bitpos, int vallen)
+get_bits_in_byte(uint8_t *from, uint16_t bitpos, uint16_t vallen)
 {
   uint16_t shifted_val;
 
@@ -109,9 +109,9 @@ get_bits_in_byte(uint8_t *from, int bitpos, int vallen)
 }
 /*---------------------------------------------------------------------------*/
 void
-get_bits(uint8_t *to, uint8_t *from, int bitpos, int vallen)
+get_bits(uint8_t *to, uint8_t *from, uint16_t bitpos, uint16_t vallen)
 {
-  int i, bits;
+  uint16_t i, bits;
   
   
   if(vallen < 8) {
@@ -170,7 +170,7 @@ header_size(const struct packetbuf_attrlist *a)
 }
 /*---------------------------------------------------------------------------*/
 static void CC_INLINE
-set_bits_in_byte(uint8_t *target, int bitpos, uint8_t val, int vallen)
+set_bits_in_byte(uint8_t *target, int bitpos, uint8_t val, uint16_t vallen)
 {
   unsigned short shifted_val;
   shifted_val = val << (8 - bitpos + 8 - vallen);
@@ -181,9 +181,9 @@ set_bits_in_byte(uint8_t *target, int bitpos, uint8_t val, int vallen)
 }
 /*---------------------------------------------------------------------------*/
 void
-set_bits(uint8_t *ptr, int bitpos, uint8_t *val, int vallen)
+set_bits(uint8_t *ptr, int bitpos, uint8_t *val, uint16_t vallen)
 {
-  int i, bits;
+  uint16_t i, bits;
 
   /*  PRINTF("set_bits %p bitpos %d, val %p len %d\n",
       ptr, bitpos, val, vallen);*/
@@ -257,8 +257,8 @@ static int
 pack_header(struct channel *c)
 {
   const struct packetbuf_attrlist *a;
-  int hdrbytesize;
-  int byteptr, bitptr, len;
+  uint16_t hdrbytesize;
+  uint16_t byteptr, bitptr, len;
   uint8_t *hdrptr;
   struct bitopt_hdr *hdr;
   
@@ -322,8 +322,8 @@ static struct channel *
 unpack_header(void)
 {
   const struct packetbuf_attrlist *a;
-  int byteptr, bitptr, len;
-  int hdrbytesize;
+  uint16_t byteptr, bitptr, len;
+  uint16_t hdrbytesize;
   uint8_t *hdrptr;
   struct bitopt_hdr *hdr;
   struct channel *c;
